Add start function and reorganize functions
