+++ index.js
@@ -1,3 +1,7 @@
+const { start } = require('../flashcards/src/game');
+
 // This is where your project starts.
 
 console.log('Your project is running...'); 
+
+start();
