+++ src/deck.js
@@ -1,0 +1,14 @@
+function createDeck(cards) {
+  const deck = cards;
+  return deck;
+}
+
+function countCards(deck) {
+  return deck.length;
+}
+
+
+module.exports = {
+  createDeck,
+  countCards,
+};
