+++ src/game.js
@@ -2,6 +2,11 @@ const data = require('./data');
 const prototypeQuestions = data.prototypeData;
 const util = require('./util');
 
+const { createDeck, countCards } = require('../src/deck');
+const { createRound } = require('../src/round');
+
+
+
 function printMessage(deck) {
   console.log(`Welcome to FlashCards! You are playing with ${countCards(deck)} cards.
   -----------------------------------------------------------------------`);
@@ -11,4 +16,13 @@ function printQuestion(round) {
   util.main(round);
 }
 
-module.exports = { printMessage, printQuestion };
+
+function start() {
+  const deck = createDeck(prototypeQuestions);
+  const round = createRound(deck);
+  printMessage(deck);
+  printQuestion(round);
+
+}
+
+module.exports = { printMessage, printQuestion, start };
